chore: move sortable list item into sortable list file
